requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.3.1
django-jazzmin==3.0.0
djangorestframework==3.15.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
redis==5.0.4
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
zope.interface==6.4.post2
pytest~=8.2.2